Refactor - Use standard directory name Fixes #13
