+++ tests/test_timer.py
@@ -11,7 +11,7 @@ people:
 
 import time
 from unittest import TestCase
-from tools.timer import Timer
+from src.timer import Timer
 
 
 class TestTimerMethods(TestCase):
